mongoDB_task_1/pyproject.toml
[project]
name = "mongodb-task-1"
version = "0.1.0"
description = ""
authors = [
    {name = "Pasha0923",email = "alexsai@i.ua"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymongo (>=4.15.5,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
